requirements.txt
-e git://github.com/slowbluecamera/slackbot.git#egg=slackbot
slackbot-eliza
tox
pytest
